user-management-service/requirements.txt
flask
ddtrace>=2.9.3
jsonify
requests
PyJWT 
boto3